artnet (#3)
Add support for using artnet devices as DMX ports. Provide a configurable timeout for how long to wait for poll responses.

Some refactoring as well - remove the available_ports method from the trait, since there's no need for this - it can just be an inherent method on each type. Provide a select_port_from function so that we do not need to repeat the polling process for each CLI port selection.